requirements.txt
loguru==0.6.0
python-dotenv==0.19.2
python-jose==3.3.0
cryptography==36.0.1
python-multipart==0.0.5
requests==2.27.1
gunicorn==20.1.0
uvicorn==0.16.0
dnspython==2.2.1
orjson==3.6.1
fastapi-utils==0.2.1
motor==2.5.1
httpx==0.22.0
fastapi==0.80.0
fastapi-jwt==0.1.8
databases
asyncpg
psycopg2-binary
celery==5.2.3
flower
passlib==1.7.4
pyperclip==1.8.2
Pillow==9.2.0
redis